feat: search info of user by username
