+++ src/Interfaces/index.ts
@@ -1,0 +1,2 @@
+export type { User } from './User'
+export type { Repository } from './Repos'
